+++ widgets/tabhost.go
@@ -6,8 +6,8 @@ import (
 
 type TabHost interface {
 	BeginExCommand(cmd string)
-	SetStatus(status string) *StatusMessage
-	SetError(err string) *StatusMessage
+	UpdateStatus()
+	SetError(err string)
 	PushStatus(text string, expiry time.Duration) *StatusMessage
 	PushError(text string) *StatusMessage
 	PushSuccess(text string) *StatusMessage